+++ src/Ares/Exception/InapplicableValidationRuleException.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * InapplicableValidationRuleException.php
+ *
+ * @author Michael Leßnau <michael.lessnau@gmail.com>
+ * @since  2019-03-30
+ */
+
+namespace Ares\Exception;
+
+use LogicException;
+
+/**
+ * Class InapplicableValidationRuleException
+ */
+class InapplicableValidationRuleException extends LogicException
+{
+}
+